requirements.txt
absl-py==1.3.0
antlr4-python3-runtime==4.9.3
attrs==22.1.0
cachetools==5.2.0
captum==0.5.0
certifi==2022.9.24
charset-normalizer==2.1.1
clearml==1.7.1
cloudpickle==2.2.0
cramjam==2.5.0
cycler==0.11.0
dask==2022.2.0
fastparquet==0.8.1
fonttools==4.37.4
fsspec==2022.10.0
furl==2.1.3
future==0.18.2
google-auth==2.13.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
hydra-core==1.2.0
hydra-submitit-launcher==1.2.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
joblib==1.2.0
jsonschema==4.16.0
kiwisolver==1.4.4
locket==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.21.6
oauthlib==3.2.2
omegaconf==2.2.3
opencv-python==4.6.0.66
orderedmultidict==1.0.1
packaging==21.3
pandas==1.3.5
partd==1.3.0
pathlib2==2.3.7.post1
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
protobuf==3.19.6
psutil==5.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
submitit==1.4.5
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
toolz==0.12.0
torch==1.12.1
torchinfo==1.7.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.9.0
